Updated aw_cognito.py file (refactored CognitoUserPool class).

+++ resources/aws_cognito.py
@@ -8,10 +8,8 @@ class CognitoUserPool:
     def __init__(self, name: str):
         self.name = name
         self._cognito_client = boto3.client('cognito-idp')
-        if self._check_exist():
-            self.arn = self.get_arn()
-        else:
-            self.arn, self.id = None, None
+        self.arn = self._get_arn()
+        self.id = self._get_id()
 
     def create_resource(self) -> None:
         """
@@ -30,16 +28,41 @@ class CognitoUserPool:
             return
         print(f'The Cognito "{self.name}" user pool exists. Using it.')
 
-    def _check_exist(self) -> bool:
+    def _get_id(self) -> str:
         """
-        Checks if Cognito user pool resource with specified name already exist.
-        Assign 'id' attribute if pool exist.
+        Returns user pool id.
         """
         for pool in self._user_pools:
             if pool['Name'] == self.name:
-                self.id = pool['Id']
-                return True
-        return False
+                return pool['Id']
+        return ''
+
+    def _remove_id(self) -> None:
+        """
+        Clear 'id' attribute.
+        """
+        self.id = ''
+
+    def _get_arn(self) -> str:
+        """
+        Returns Cognito user pool ARN.
+        """
+        try:
+            return self._cognito_client.describe_user_pool(UserPoolId=self.id)['UserPool']['Arn']
+        except AttributeError:
+            return ''
+
+    def _remove_arn(self) -> None:
+        """
+        Clear 'arn' attribute.
+        """
+        self.arn = ''
+
+    def _check_exist(self) -> bool:
+        """
+        Checks if Cognito user pool resource with specified name already exist.
+        """
+        return True if self.id else False
 
     @property
     def _user_pools(self) -> list:
@@ -47,12 +70,6 @@ class CognitoUserPool:
         Returns list of already created Cognito user pools.
         """
         return self._cognito_client.list_user_pools(MaxResults=60)['UserPools']
-
-    def get_arn(self) -> str:
-        """
-        Returns Cognito user pool ARN.
-        """
-        return self._cognito_client.describe_user_pool(UserPoolId=self.id)['UserPool']['Arn']
 
     @property
     def domain(self) -> str:
@@ -69,8 +86,16 @@ class CognitoUserPool:
         """
         Deletes Cognito user pool in AWS cloud.
         """
-        self._cognito_client.delete_user_pool(UserPoolId=self.id)
-        print(f'The Cognito "{self.name}" user pool has been deleted')
+        if self._check_exist():
+            try:
+                self._cognito_client.delete_user_pool(UserPoolId=self.id)
+                print(f'The Cognito user pool "{self.name}" has been deleted')
+                self._remove_id()
+                self._remove_arn()
+            except self._cognito_client.exceptions.InvalidParameterException as err:
+                print(err.response['Error']['Message'])
+            return
+        print(f'The Cognito user pool "{self.name}" does not exist')
 
 
 class CognitoResourceServer:
@@ -188,7 +213,7 @@ class CognitoUserPoolClient:
         self.user_pool_id = user_pool_id
         self.scopes = scopes
         self._cognito_client = boto3.client('cognito-idp')
-        self.id = self._get_client_id()
+        self.id = self._get_id()
 
     def create_resource(self) -> None:
         """
@@ -215,7 +240,7 @@ class CognitoUserPoolClient:
         return self._cognito_client.list_user_pool_clients(MaxResults=60,
                                                            UserPoolId=self.user_pool_id)['UserPoolClients']
 
-    def _get_client_id(self) -> str:
+    def _get_id(self) -> str:
         """
         Returns user pool client id.
         """
@@ -224,7 +249,7 @@ class CognitoUserPoolClient:
                 return client['ClientId']
         return ''
 
-    def _remove_client_id(self) -> None:
+    def _remove_id(self) -> None:
         """
         Clear 'id' attribute.
         """
@@ -246,7 +271,7 @@ class CognitoUserPoolClient:
             self._cognito_client.delete_user_pool_client(UserPoolId=self.user_pool_id,
                                                          ClientId=self.id)
             print(f'The Cognito user pool client "{self.name}" has been deleted')
-            self._remove_client_id()
+            self._remove_id()
             return
         print(f'The Cognito user pool client "{self.name}" does not exist')
